fix: don't return gasUsed when estimateGas went wrong
